mio_band_update/requirements.txt
pyserial==3.5
stm32loader==0.5.1
esptool==4.2